+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 3.13.0
 * Support InputStreams as bodies.
 * Support ProgressMonitors for InputStream bodies
+* Unirest-Mocks now includes a MockResponse<T> and a MockConfig for use independent of the MockClient
 
 ## 3.12.0
 * Bump GSON to 2.8.8